Win32: build 2.5 tagged releases from git instead of svn
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@22979 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ packaging/win32/build_tags.sh
@@ -56,6 +56,17 @@ mv -f ${tagfile}.new ${tagfile}
 for tag_rev in $tags ; do
   tag=${tag_rev#*/}
   tag=${tag%/*}
+  
+  # From 2.5 and up, tags are built from git
+  tag_major=${tag%%.*}
+  tag_tmp=${tag#*.}
+  tag_minor=${tag_tmp%%.*}
+  major_minor=$(( $tag_major*100 + $tag_minor ))
+  if (( $major_minor >= 205 ))
+  then
+     continue
+  fi
+  
   tagbasedir=/c/soft/gnucash-${tag}
   tagdir=${tagbasedir}/gnucash
   rm -fr $tagbasedir
